rsa_services-prod_build/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
pandas==1.5.3
SQLAlchemy==1.4.46
geoalchemy2==0.13.0
Pillow==10.0.1
PyJWT==2.8.1
gunicorn==21.2.0
shapely==1.8.5
geopandas==0.12.2
fiona==1.9.5
